+++ saleor/graphql/checkout/mutations/utils.py
@@ -38,6 +38,8 @@ ERROR_DOES_NOT_SHIP = "This checkout doesn't need shipping"
 
 @dataclass
 class CheckoutLineData:
+    variant_id: Optional[str] = None
+    line_id: Optional[str] = None
     quantity: int = 0
     quantity_to_update: bool = False
     custom_price: Optional[Decimal] = None
@@ -94,6 +96,26 @@ def update_checkout_shipping_method_if_invalid(
 
     if not is_valid:
         clear_delivery_method(checkout_info)
+
+
+def get_variants_and_total_quantities(
+    variants, lines_data, quantity_to_update_check=False
+):
+    variants_total_quantity_map = defaultdict(int)
+    mapped_data = defaultdict(int)
+
+    if quantity_to_update_check:
+        lines_data = filter(lambda d: d.quantity_to_update, lines_data)
+
+    for data in lines_data:
+        mapped_data[data.variant_id] += data.quantity
+
+    for variant in variants:
+        quantity = mapped_data.get(str(variant.id), None)
+        if quantity is not None:
+            variants_total_quantity_map[variant] += quantity
+
+    return variants_total_quantity_map.keys(), variants_total_quantity_map.values()
 
 
 def check_lines_quantity(
@@ -273,24 +295,105 @@ def get_checkout(
     return checkout
 
 
-def group_quantity_and_custom_prices_by_variants(
-    lines: List[Dict[str, Any]]
+def group_lines_input_on_add(
+    lines: List[Dict[str, Any]], existing_lines_info=None
 ) -> List[CheckoutLineData]:
-    variant_checkout_line_data_map: Dict[str, CheckoutLineData] = defaultdict(
-        CheckoutLineData
-    )
+    """Return list od CheckoutLineData objects.
+
+    Lines data provided in CheckoutLineInput will be grouped depending on
+    provided parameters.
+    """
+    grouped_checkout_lines_data: List[CheckoutLineData] = []
+    checkout_lines_data_map: Dict[str, CheckoutLineData] = defaultdict(CheckoutLineData)
 
     for line in lines:
         variant_id = cast(str, line.get("variant_id"))
-        line_data = variant_checkout_line_data_map[variant_id]
+        force_new_line = line.get("force_new_line")
+        _, variant_db_id = graphene.Node.from_global_id(variant_id)
+
+        if force_new_line:
+            line_data = CheckoutLineData(variant_id=variant_db_id)
+            grouped_checkout_lines_data.append(line_data)
+        else:
+            _, variant_db_id = graphene.Node.from_global_id(variant_id)
+
+            try:
+                line_db_id = find_line_id_when_variant_parameter_used(
+                    variant_db_id, existing_lines_info
+                )
+
+                if not line_db_id:
+                    line_data = checkout_lines_data_map[variant_db_id]
+                    line_data.variant_id = variant_db_id
+                else:
+                    line_data = checkout_lines_data_map[line_db_id]
+                    line_data.line_id = line_db_id
+                    line_data.variant_id = find_variant_id_when_line_parameter_used(
+                        line_db_id, existing_lines_info
+                    )
+
+            # when variant already exist in multiple lines then create a new line
+            except ValidationError:
+                line_data = CheckoutLineData(variant_id=variant_db_id)
+                grouped_checkout_lines_data.append(line_data)
+
         if (quantity := line.get("quantity")) is not None:
             line_data.quantity += quantity
             line_data.quantity_to_update = True
+
         if "price" in line:
             line_data.custom_price = line["price"]
             line_data.custom_price_to_update = True
 
-    return list(variant_checkout_line_data_map.values())
+    grouped_checkout_lines_data += list(checkout_lines_data_map.values())
+    return grouped_checkout_lines_data
+
+
+def group_lines_input_data_on_update(
+    lines: List[Dict[str, Any]], existing_lines_info=None
+) -> List[CheckoutLineData]:
+    """Return list od CheckoutLineData objects.
+
+    This function is used in CheckoutLinesUpdate mutation.
+    Lines data provided in CheckoutLineUpdateInput will be grouped depending on
+    provided parameters.
+    """
+    grouped_checkout_lines_data: List[CheckoutLineData] = []
+    checkout_lines_data_map: Dict[str, CheckoutLineData] = defaultdict(CheckoutLineData)
+
+    for line in lines:
+        variant_id = cast(str, line.get("variant_id"))
+        line_id = cast(str, line.get("line_id"))
+
+        if line_id:
+            _, line_db_id = graphene.Node.from_global_id(line_id)
+
+        if variant_id:
+            _, variant_db_id = graphene.Node.from_global_id(variant_id)
+            line_db_id = find_line_id_when_variant_parameter_used(
+                variant_db_id, existing_lines_info
+            )
+
+        if not line_db_id:
+            line_data = checkout_lines_data_map[variant_db_id]
+            line_data.variant_id = variant_db_id
+        else:
+            line_data = checkout_lines_data_map[line_db_id]
+            line_data.line_id = line_db_id
+            line_data.variant_id = find_variant_id_when_line_parameter_used(
+                line_db_id, existing_lines_info
+            )
+
+        if (quantity := line.get("quantity")) is not None:
+            line_data.quantity += quantity
+            line_data.quantity_to_update = True
+
+        if "price" in line:
+            line_data.custom_price = line["price"]
+            line_data.custom_price_to_update = True
+
+    grouped_checkout_lines_data += list(checkout_lines_data_map.values())
+    return grouped_checkout_lines_data
 
 
 def check_permissions_for_custom_prices(app, lines):
@@ -303,3 +406,49 @@ def check_permissions_for_custom_prices(app, lines):
         not app or not app.has_perm(CheckoutPermissions.HANDLE_CHECKOUTS)
     ):
         raise PermissionDenied(permissions=[CheckoutPermissions.HANDLE_CHECKOUTS])
+
+
+def find_line_id_when_variant_parameter_used(
+    variant_db_id: str, lines_info: List[CheckoutLineInfo]
+):
+    """Return line id when variantId parameter was used.
+
+    If variant exists in multiple lines error will be returned.
+    """
+    if not lines_info:
+        return
+
+    line_info = list(filter(lambda x: (x.variant.pk == int(variant_db_id)), lines_info))
+
+    if not line_info:
+        return
+
+    # if same variant occur in multiple lines `lineId` parameter have to be used
+    if len(line_info) > 1:
+        message = (
+            "Variant occurs in multiple lines. Use `lineId` instead " "of `variantId`."
+        )
+        variant_global_id = graphene.Node.to_global_id("ProductVariant", variant_db_id)
+
+        raise ValidationError(
+            {
+                "variantId": ValidationError(
+                    message=message,
+                    code=CheckoutErrorCode.INVALID.value,
+                    params={"variants": [variant_global_id]},
+                )
+            }
+        )
+
+    return str(line_info[0].line.id)
+
+
+def find_variant_id_when_line_parameter_used(
+    line_db_id: str, lines_info: List[CheckoutLineInfo]
+):
+    """Return variant id when lineId parameter was used."""
+    if not lines_info:
+        return
+
+    line_info = list(filter(lambda x: (str(x.line.pk) == line_db_id), lines_info))
+    return str(line_info[0].line.variant_id)